Implemented the recommender system in 2 different ways
I organized the repository to be more organized. There is 2 different files called MovieRecommendSystem and SimpledMovieRecommend that have been tested through a database that has 1000 entries that can be found in database folder. The readme in code folder gives a rough description of code files and the code files have been commented to give a full idea of what the code should be doing